[HumanBenchmark][Backend[SocialGraph] Refactor Social graph backend further
